Julia_Hashcode2014

Documentation for Julia_Hashcode2014.

Julia_Hashcode2014.AdjacencyGraph

Julia_Hashcode2014.TreeNode

Julia_Hashcode2014.dijkstra

Julia_Hashcode2014.edge_time

Julia_Hashcode2014.edge_weight

Julia_Hashcode2014.fandown_greedy

Julia_Hashcode2014.find_bound

Julia_Hashcode2014.find_nesw

Julia_Hashcode2014.get_valid_candidates

Julia_Hashcode2014.greedy

Julia_Hashcode2014.greedy_lookahead

Julia_Hashcode2014.greedy_lookahead_dijkstras_fandown

Julia_Hashcode2014.lookahead_tree

Julia_Hashcode2014.lookahead_tree_bounded

Julia_Hashcode2014.node_metric

Julia_Hashcode2014.outneighbors

Julia_Hashcode2014.recalculate_times

Julia_Hashcode2014.remove_dups

Julia_Hashcode2014.spath

Julia_Hashcode2014.AdjacencyGraph — Type

AdjacencyGraph

Store a city in the form of an weighted adjacency graph. Weights are based on length.

Fields

outneighbors::Vector{Vector{Int}}: junctions

weights::Vector{Vector{Float64}}: weights based on length for each street.

times::Dict{Tuple{Int,Int},Float64}: the time it takes to travel a street.

Julia_Hashcode2014.TreeNode — Type

TreeNode

Stores nodes in a graph. Each node only has 1 parent node. Used for backtracking BFS lookahead trees.

Fields

parent::Union{TreeNode, Nothing}: The parent node/junction.

junction::Int64: The junction index.

time_elapsed::Float64: the total time that has elapsed to reach this junction from the start.

distance_traveled::Float64: the total distrance traveled to reach this junction from the start.

path_visited::Set{Tuple{Int64, Int64}}: the (u, v) paths taken to reach this junction from the start.

Julia_Hashcode2014.dijkstra — Method

dijkstra(g, s)

Finds the shortest path (in terms of time) from the source to all other points in the graph g and the parents list to reconstruct the shortest path. Returns a tuple (shortest_path, parents).

Parameters

g: The city as a AdjacencyGraph.

s: The source junction index.

Julia_Hashcode2014.edge_time — Method

edge_time(g, u, v)

Returns the time it takes to traverse the path (u, v) based on an AdjacencyGraph g.

Parameters

g::AdjacencyGraph: The adjacency graph.

u: a node u.

v: a node v.

Julia_Hashcode2014.edge_weight — Method

edge_weight(g, u, v)

Returns the distance that the path (u, v) will traverse based on an AdjacencyGraph g.

Parameters

g::AdjacencyGraph: The adjacency graph.

u: a node u.

v: a node v.

Julia_Hashcode2014.fandown_greedy — Function

fandown_greedy(city, n_lookahead = 15, seq_steps = 7, init_iter = 100, diff_rand = 5)

A greedy algorithm that uses BFS to lookahead from the start node throughout the graph while obeying the time constraints. At the start, has all cars "fandown" to reach a more center position in the city. Uses a distance metric to select best itinerary.

Takes around 15 seconds to run (with default parameters) and obtains a score of around 1.35-1.47 million.

Parameters

city: The city (as a City from HashCode2014)

n_lookahead: The number of BFS levels to lookahead.

seq_steps: the number of steps to take for each car per round.

init_iter: number of iterations to fandown the cars in the beginning

diff_rand: probability must taking the same path

Julia_Hashcode2014.find_bound — Function

find_bound(city, nb_cars = 8, time = 54000)

Finds the maximum distance bound for a city based on the number of cars and the time constraint. Sorts the streets by speed, and takes turns sending cars down each successive street until the time constraint is met.

Parameters

city::City: the city.

nb_cars::Int: the number of cars.

time::Float64: the time constraint in seconds.

Julia_Hashcode2014.find_nesw — Method

find_nesw(city)

Finds the most north, east, south, and west points of a city by using the min/max latitude and longitude. Returns result as a vector [N, E, S, W] with elements as junction indices.

Parameters

city::City: The city (as a City from HashCode2014)

Julia_Hashcode2014.get_valid_candidates — Method

get_valid_candidates(graph, current, candidates, time_elapsed, time_remaining)

Returns a list of valid neighboring junctions from the current junction. A neighbor is valid if the time it requires to travel there fits within the time requirement.

Parameters

graph::AdjacencyGraph: The AdjacencyGraph representing the city.

current: the current junction index.

candidates: neighboring junctions.

time_elapsed: the time elapsed so far.

time_remaining: the remaining time for the car to travel.

Julia_Hashcode2014.greedy — Method

greedy(city)

Implements a greedy algorithm for routing. Cars will take turns going down one street at a time based on the follow criteria:

Check if the neighboring node has been visited. If not, take the maximum distance path while obeying time constraints.

If a neighboring node has been visited, check the next node.

If all neighboring nodes have been visited, choose a random path to go down while obeying the time constraint.

Takes around 1-2 seconds to run and obtains a score of around 1.1-1.25 million.

Parameters

city::City: The city (as a City from HashCode2014)

Julia_Hashcode2014.greedy_lookahead — Function

greedy_lookahead(city, n_lookahead, seq_steps)

A greedy algorithm that uses BFS to lookahead from the start node throughout the graph while obeying the time constraints. Uses a distance metric to select best itinerary.

Takes around 20 seconds to run (with default parameters) and obtains a score of around 1.3-1.45 million.

Parameters

city::City: The city (as a City from HashCode2014)

n_lookahead: The number of BFS levels to lookahead.

seq_steps: the number of steps to take for each car per round.

Julia_Hashcode2014.greedy_lookahead_dijkstras_fandown — Function

function greedy_lookahead_dijkstras_fandown(city, n_lookahead = 15, seq_steps = 15)

A greedy algorithm that uses BFS to lookahead from the start node throughout the graph while obeying the time constraints. Sends 4 cars to the N/E/S/W corners using Dijkstra's and uses fandown for the remaining cars. Uses lookaheadtreebounded as a lookahead algorithm and uses a speed metric to select best itinerary. Removes duplicates and repeats the lookahead algorithm.

Takes around 30 seconds to run (with default parameters) and obtains a score of around 1.65-1.8 million.

Parameters

city::City: The city (as a City from HashCode2014)

n_lookahead: The number of BFS levels to lookahead.

seq_steps: the number of steps to take for each car per round.

Julia_Hashcode2014.lookahead_tree — Method

lookahead_tree(graph, start, visited, n, time_remaining)

Uses BFS to lookahead from the start node throughout the graph while obeying the time constraints. Uses a distance metric to select best itinerary. Any paths already traveled are heavily discounted in the metric. Returns a (best distance, best itinerary) tuple.

Parameters

graph::AdjacencyGraph: The AdjacencyGraph representing the city.

start: The starting junction index.

visited: the (u, v) paths that have already been taken by any cars.

n: the number of levels to iterate for BFS.

time_remaining: the remaining time for the car to travel.

Julia_Hashcode2014.lookahead_tree_bounded — Method

lookahead_tree_bounded(graph, start, visited, n, time_remaining, city, lat_min, lat_max, long_min, long_max)

Uses BFS to lookahead from the start node throughout the graph while obeying the time constraints. Uses a speed metric to select best itinerary. Any paths already traveled are heavily discounted in the metric. Any junctions outside the min/max latitude/longitude are also heavily discounted. Returns a (best distance, best itinerary) tuple.

Parameters

graph::AdjacencyGraph: The AdjacencyGraph representing the city.

start: The starting junction index.

visited: the (u, v) paths that have already been taken by any cars.

n: the number of levels to iterate for BFS.

time_remaining: the remaining time for the car to travel.

city: the city.

lat_min: the minimum latitude.

lat_max: the maximum latitude.

long_min: the minimum longitude.

long_max: the maximum longitude.

Julia_Hashcode2014.node_metric — Method

node_metric(node)

Returns the metric for the current node. Currently uses distance/time as the metric.

Parameters

node::TreeNode: the node in the BFS lookahead tree.

Julia_Hashcode2014.outneighbors — Method

outneighbors(g, u)

Returns the outneighbors of node u based on an AdjacencyGraph g.

Parameters

g::AdjacencyGraph: The adjacency graph.

u: a node u.

Julia_Hashcode2014.recalculate_times — Method

recalculate_times(graph, itineraries)

Calculates the amount of time needed to fulfill the itineraries.

Parameters

graph::AdjacencyGraph: The AdjacencyGraph representing the city.

itinerary::Vector{Vector{Int64}}: the vector itinerary of a car.

Julia_Hashcode2014.remove_dups — Method

remove_dups(itinerary)

Returns the itinerary after removing duplicates. Removed duplicates that are in the form [...,1,2,1,2,...] -> [...,1,2,...].

Parameters

itinerary::Vector{Int64}: the vector itinerary of a car.

Julia_Hashcode2014.spath — Method

spath(x, dists, source, parents)

Returns the shortest path (in terms of time) from the source to the junction x.

Parameters

x: The target junction index.

dists: The shortest paths in terms of time from dijkstra.

source: The source junction index.

parents: The parents array from dijkstra.