Julia_Hashcode2014

Documentation for Julia_Hashcode2014.

Julia_Hashcode2014.AdjacencyGraph

Julia_Hashcode2014.create_graph

Julia_Hashcode2014.edge_time

Julia_Hashcode2014.edge_weight

Julia_Hashcode2014.greedy

Julia_Hashcode2014.outneighbors

Julia_Hashcode2014.AdjacencyGraph — Type

AdjacencyGraph

Store a city in the form of an weighted adjacency graph. Weights are based on length.

Fields

outneighbors::Vector{Vector{Int}}: junctions

weights::Vector{Vector{Float64}}: weights based on length for each street.

times::Dict{Tuple{Int,Int},Float64}: the time it takes to travel a street.

Julia_Hashcode2014.create_graph — Method

Create an [`AdjacencyGraph`](@ref) based on a city.

Julia_Hashcode2014.edge_time — Method

edge_time(g, u, v)

Returns the time it takes to traverse the path (u, v) based on an AdjacencyGraph g.

Julia_Hashcode2014.edge_weight — Method

edge_weight(g, u, v)

Returns the distance that the path (u, v) will traverse based on an AdjacencyGraph g.

Julia_Hashcode2014.greedy — Method

greedy(city)

Implements a greedy algorithm for routing. Cars will take turns going down one street at a time based on the follow criteria:

Check if the neighboring node has been visited. If not, take the maximum distance path while obeying time constraints.

If a neighboring node has been visited, check the next node.

If all neighboring nodes have been visited, choose a random path to go down while obeying the time constraint.

Julia_Hashcode2014.outneighbors — Method

outneighbors(g, u)

Returns the outneighbors of node u based on an AdjacencyGraph g.