Julia_Hashcode2014

Documentation for Julia_Hashcode2014.

Julia_Hashcode2014.AdjacencyGraph

Julia_Hashcode2014.City

Julia_Hashcode2014.Junction

Julia_Hashcode2014.Solution

Julia_Hashcode2014.Street

Julia_Hashcode2014.edge_time

Julia_Hashcode2014.edge_weight

Julia_Hashcode2014.get_street_end

Julia_Hashcode2014.greedy

Julia_Hashcode2014.is_feasible

Julia_Hashcode2014.is_street

Julia_Hashcode2014.is_street_start

Julia_Hashcode2014.outneighbors

Julia_Hashcode2014.plot_streets

Julia_Hashcode2014.read_city

Julia_Hashcode2014.read_solution

Julia_Hashcode2014.total_distance

Julia_Hashcode2014.write_city

Julia_Hashcode2014.write_solution

Julia_Hashcode2014.AdjacencyGraph — Type

AdjacencyGraph

Store a city in the form of an weighted adjacency graph. Weights are based on length.

Fields

outneighbors::Vector{Vector{Int}}: junctions

weights::Vector{Vector{Float64}}: weights based on length for each street.

times::Dict{Tuple{Int,Int},Float64}: the time it takes to travel a street.

Julia_Hashcode2014.City — Type

City

Store a city made of Junctions and Streets, along with additional instance parameters.

Fields

total_duration::Int: total time allotted for the car itineraries (in seconds)

nb_cars::Int: number of cars in the fleet

starting_junction::Int: junction at which all the cars are located initially

junctions::Vector{Junction}: list of junctions

streets::Vector{Street}: list of streets

Julia_Hashcode2014.Junction — Type

Junction

Store a city junction.

Fields

latitude::Float64: latitude (in decimal degrees)

longitude::Float64: longitude (in decimal degrees)

Julia_Hashcode2014.Solution — Type

Solution

Store a set of itineraries, one for each car.

Fields

itineraries::Vector{Vector{Int}}: each itinerary is a vector of junction indices

Julia_Hashcode2014.Street — Type

Street Store an edge between two Junctions.

Fields

endpointA::Int: index of the first junction

endpointB::Int: index of the second junction

bidirectional::Bool: whether B -> A is allowed

duration::Int: time cost of traversing the street (in seconds)

distance::Int: length of the street (in meters)

Julia_Hashcode2014.edge_time — Method

edge_time(g, u, v)

Returns the time it takes to traverse the path (u, v) based on an AdjacencyGraph g.

Julia_Hashcode2014.edge_weight — Method

edge_weight(g, u, v)

Returns the distance that the path (u, v) will traverse based on an AdjacencyGraph g.

Julia_Hashcode2014.get_street_end — Method

get_street_end(i, street)

Retrieve the arrival endpoint of street when it starts at junction i.

Julia_Hashcode2014.greedy — Method

greedy(city)

Implements a greedy algorithm for routing. Cars will take turns going down one street at a time based on the follow criteria:

Check if the neighboring node has been visited. If not, take the maximum distance path while obeying time constraints.

If a neighboring node has been visited, check the next node.

If all neighboring nodes have been visited, choose a random path to go down while obeying the time constraint.

Julia_Hashcode2014.is_feasible — Method

is_feasible(solution, city[; verbose=false])

Check if solution satisfies the constraints for the instance defined by city. The following criteria are considered (taken from the problem statement):

the number of itineraries has to match the number of cars of city

the first junction of each itinerary has to be the starting junction of city

for each consecutive pair of junctions on an itinerary, a street connecting these junctions has to exist in city (if the street is one directional, it has to be traversed in the correct direction)

the duration of each itinerary has to be lower or equal to the total duration of city

Julia_Hashcode2014.is_street — Method

is_street(i, j, street)

Check if the trip from junction i to junction j corresponds to a valid direction of street.

Julia_Hashcode2014.is_street_start — Method

is_street_start(i, street)

Check if junction i corresponds to a valid starting point of street.

Julia_Hashcode2014.outneighbors — Method

outneighbors(g, u)

Returns the outneighbors of node u based on an AdjacencyGraph g.

Julia_Hashcode2014.plot_streets — Function

plot_streets(city, solution=nothing; path=nothing)

Plot a City and an optional Solution using the Python library folium, save the result as an HTML file at path.

Julia_Hashcode2014.read_city — Function

read_city(path)

Read and parse a City from a file located at path. The default path is an artifact containing the official challenge data.

Julia_Hashcode2014.read_solution — Method

read_solution(solution, path)

Read and parse a Solution from a file located at path.

Julia_Hashcode2014.total_distance — Method

total_distance(solution, city)

Compute the total distance of all itineraries in solution based on the street data from city. Streets visited several times are only counted once.

Julia_Hashcode2014.write_city — Method

write_city(city, path)

Write a City to a file located at path.

Julia_Hashcode2014.write_solution — Method

write_solution(solution, path)

Write a Solution to a file located at path.